+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import PokemonsPage from './components/PokemonsPage';
+import PokemonsPage from './components/PokemonsPage/PokemonsPage';
 import ErrorBoundary from './components/ErrorBoundary';
 
 export default class App extends Component {
